ruby updates
This got close to working when added piecemeal; let's
see.

+++ ubuntu/ruby/setup.sh
@@ -1,38 +1,50 @@
 # https://www.ruby-lang.org/en/documentation/installation/#building-from-source
 # https://github.com/ruby/ruby/blob/master/doc/contributing/building_ruby.md
 #
+source ~/scripts/util/detect-installer.sh
+echo "Linux distribution: $BVT_DISTRO"
+echo "Selected installer: $BVT_INSTALLER"
+echo "***** BEGIN RUBY SETUP SCRIPT $(date) *****"
+
+PKGMGR=$BVT_INSTALLER
+
 # Some of these may still be unneeded -- be mindful of parsimony
 # It seems that apt (Ubuntu) repositories package the same libraries under slightly
 # different names, perhaps owing to dev versus devel
 #
 # See packages.ubuntu.com
 pushd ~
-PKGMGR=apt
-sudo ${PKGMGR} install -y gcc
-# sudo ${PKGMGR} install -y libyaml-2-0
-sudo ${PKGMGR} install -y libyaml-dev
-sudo ${PKGMGR} install -y zlib1g-dev
-sudo ${PKGMGR} install -y autoconf
-sudo ${PKGMGR} install -y bison
-sudo ${PKGMGR} install -y gperf
-sudo ${PKGMGR} install -y libffi-dev
-sudo ${PKGMGR} install -y openssl
-sudo ${PKGMGR} install -y readline-common
-# sudo ${PKGMGR} install -y ruby	# according to official ruby instructions
-sudo ${PKGMGR} install -y ruby-dev
+#PKGMGR=apt
+sudo ${BVT_INSTALLER} update
+sudo ${BVT_INSTALLER} install -y gcc
+# sudo ${BVT_INSTALLER} install -y libyaml-2-0
+sudo ${BVT_INSTALLER} install -y libyaml-dev
+sudo ${BVT_INSTALLER} install -y zlib1g-dev
+sudo ${BVT_INSTALLER} install -y autoconf
+sudo ${BVT_INSTALLER} install -y bison
+sudo ${BVT_INSTALLER} install -y gperf  # Ubuntu: unable to find
+sudo ${BVT_INSTALLER} install -y libffi-dev
+sudo ${BVT_INSTALLER} install -y openssl
+sudo ${BVT_INSTALLER} install -y readline-common
+# sudo ${BVT_INSTALLER} install -y ruby	# according to official ruby instructions
+sudo ${BVT_INSTALLER} install -y ruby-dev
 
 sudo gem install bundler
 git clone --depth 1 https://github.com/ruby/ruby.git 
 	pushd ruby
 	./autogen.sh
-	mv build build-backup
+	[ -f build ] && mv build build-backup
 	mkdir build
 	popd
 popd
 
 
+# Phase 1 comment:
 # Although it seemed like a longshot this actually made everything but psych appear to work.
 # Let's try installing rvm too just for good measure.
+# Phase 2 -- it can't find rvm on the path after that. And since the build fails
+# badly this seems to be a problem.
+
 sudo gem install rvm
 #
 # https://github.com/rvm/rvm/issues/5069
